refactor: separate types declaration to different files
