+++ backend/src/types/db/index.ts
@@ -1,0 +1,12 @@
+export * from "./client.js";
+export * from "./user.js";
+export * from "./ticket.js";
+export * from "./equipment.js";
+export * from "./category.js";
+export * from "./ticketMessage.js";
+export * from "./aiResponse.js";
+export * from "./aiFeedback.js";
+export * from "./resolvedCase.js";
+export * from "./refreshToken.js";
+export * from "./attachment.js";
+export * from "./knowledgeArticle.js";
