+++ backend/src/types/db/user.ts
@@ -1,0 +1,19 @@
+import type { Role } from "../misc/roles.js";
+
+/**
+ * Represents the full structure of the 'users' table.
+ */
+export interface UserDB {
+  id: string; // UUID
+  client_id: string;
+  email: string;
+  password_hash: string;
+  role: Role;
+  name: string;
+  created_at: Date;
+}
+
+/**
+ * Type for creating a new user, based on the full DB entity.
+ */
+export type NewUser = Omit<UserDB, "id" | "created_at">;
