+++ backend/src/types/index.ts
@@ -1,0 +1,9 @@
+// types/index.ts
+
+/**
+ * This file exports the types used across the application, including database entities and data transfer objects (DTOs).
+ *  */
+
+export * from "./db/index.js";
+export * from "./auth/jwt.js";
+export * from "./misc/index.js";
