+++ backend/src/types/misc/index.ts
@@ -1,0 +1,2 @@
+export * from "./roles.js";
+export * from "./statuses.js";
